Move dialog to its own file
