+++ src/PointDataConversionPlugin.h
@@ -136,26 +136,3 @@ private:
     mv::gui::DecimalAction _arcSinFactorAction;
     mv::gui::SlidersAction _arcSinFactorsAction;
 };
-
-/**
- * Helper dialog to set conversion options
- *
- * @author Alex Vieth
- */
-class ConversionDialog : public QDialog
-{
-    Q_OBJECT
-public:
-    explicit ConversionDialog(QWidget* parent, mv::gui::ToggleAction* sameFactorAction, mv::gui::DecimalAction* arcSinFactorAction, mv::gui::SlidersAction* arcSinFactorsAction);
-
-signals:
-    void closeDialog(bool onlyIndices);
-
-private slots:
-    void closeDialogAction() {
-        emit QDialog::accept();
-    }
-
-private:
-    mv::gui::TriggerAction _conversionButton;
-};
